requirements.txt
pytest
poetry
pandas
biopython
numpy